trying to get story data on page
