+++ ambrosia_cafe/src/components/storySection.js
@@ -1,0 +1,53 @@
+import React from 'react'
+import store from '../store'
+import {getMenu} from '../api/menu'
+
+const styles = { 
+  menuCategory:{
+    fontSize: 22,
+    fontFamily: 'Rock Salt',
+    border: 'solid black 2px',
+    width: '100%',
+    display: 'inline-block',
+    borderWidth: '0 0 2px 0',
+    paddingBottom: 4,
+    textAlign: 'left'
+  },
+  storyText:{
+  	width: 840,
+    height: 300,
+    display:'inline-block',
+    padding:'20px 25px 10px 25px',
+  }
+}
+
+class StorySection extends React.Component {
+  constructor() {
+    super()
+    this.state = { story: "" }
+  }
+  componentWillMount() {
+      this.unsubscribe = store.subscribe(()=>{
+          const appState = store.getState().menuReducer.menu.story.text
+    
+      this.setState({
+            menu: appState
+          })
+      })
+      getMenu()
+    }
+    componentWillUnmount() {
+      this.unsubscribe()
+    }
+  render() {
+    return (
+      <div>
+        <div className="leftPhotos">IMAGES STACK IN HERE</div>
+      		<div style={styles.storyText}>{this.state.story} </div>
+      	<div className="rightPhotos">IMAGES STACK IN HERE</div>
+      </div>
+    )
+  }
+}
+
+export default StorySection
